requirements.txt
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-openai==0.1.23

langchain-google-genai==1.0.10
langchain-pinecone==0.2.13

sentence-transformers==2.6.1
pypdf==4.2.0
python-dotenv==1.0.1

pinecone==5.0.1

flask==3.0.3

-e .
